requirements.txt
# Web Framework & API
Flask
requests
Flask-HTTPAuth
gunicorn

# AI/LLM Tools
openai
transformers
torch
accelerate
safetensors
outlines[transformers]  # This will install outlines with transformers support
outlines[openai]       # This will add OpenAI support

# Integration & Data
notion-client
python-dotenv
pydantic
jsonschema

# Document Processing
python-docx
PyPDF2
pylatex
jinja2

# Utilities
python-dateutil
beautifulsoup4
Pillow
rich
typer
loguru
python-dateutil

# Development Tools
pytest
pytest-cov
black
ruff
responses